new demo page, better scoring calc
