+++ docs/demo/demo.css
@@ -1,0 +1,33 @@
+body{
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    flex-wrap: wrap;
+    gap:80px;
+    justify-content: center;
+    font-family: system-ui, -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
+    padding-bottom: 300px;
+
+}
+
+.dvsc_schowcase{
+    display: flex;
+    flex-direction: column;
+    flex-wrap: wrap;
+    justify-content: center;
+    align-items: center;
+    gap: 50px;
+}
+
+#config_obj{
+    width: 80vw;
+    height: 30vh;
+}
+
+.description{
+    max-width: 50vw;
+}
+
+.guide{
+    max-width: 70vw;
+}
