+++ docs/demo/style.css
@@ -1,0 +1,137 @@
+
+div.dynamic_visual_scoring_calculator#no1 {
+    display: flex;
+    flex-direction: row;
+    flex-wrap: wrap;
+    justify-content: center;
+    justify-items: center;
+    align-items: center;
+    align-content: center;
+    gap: 100px;
+    font-family: system-ui, -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
+}
+
+div.dynamic_visual_scoring_calculator#no1>div.dvsc_settings>table>tbody{
+    display: flex;
+    flex-direction: column;
+    flex-wrap: wrap;
+    align-content: flex-start;
+    justify-content: flex-start;
+    gap: 17px;
+    align-items: stretch;
+}
+
+div.dynamic_visual_scoring_calculator#no1>div.dvsc_settings>table>thead>tr {
+    border-bottom: 1px solid #64646480;
+    padding: 8px 0;
+}
+
+div.dynamic_visual_scoring_calculator#no1>div.dvsc_settings>table>thead>tr,
+div.dynamic_visual_scoring_calculator#no1>div.dvsc_settings>table>tbody>tr {
+    display: grid;
+    grid-gap: 32px;
+    grid-auto-flow: column;
+    grid-template-columns: 75px 62px 197px 164px 190px 110px;
+    justify-content: space-between;
+}
+
+div.dynamic_visual_scoring_calculator#no1>div.dvsc_settings>table>tbody>tr>td {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    gap: 12px;
+    flex-wrap: nowrap;
+    text-align: end;
+}
+
+div.dynamic_visual_scoring_calculator#no1>div.dvsc_settings>table>thead>tr>th{
+    font-weight: 500;
+    opacity: 0.5;
+}
+
+div.dynamic_visual_scoring_calculator#no1>div.dvsc_settings>table>thead>tr>th:nth-child(1) {
+    text-align: left;
+}
+
+div.dynamic_visual_scoring_calculator#no1>div.dvsc_settings>table>tbody>tr>td:nth-child(2) {
+    justify-content: center;
+}
+
+
+div.dynamic_visual_scoring_calculator#no1>div.dvsc_settings>table>tbody>tr>td:nth-child(6){
+    justify-content: flex-end;
+}
+
+div.dynamic_visual_scoring_calculator#no1>div.dvsc_settings>table>tbody>tr>th {
+    text-align: left;
+    justify-content: flex-start;
+    align-items: center;
+    display: flex;
+    flex-direction: row;
+    flex-wrap: nowrap;
+    align-content: center;
+}
+
+
+div.dynamic_visual_scoring_calculator#no1>div.dvsc_settings>table>tbody>tr>td>.dvsc_tab,
+div.dynamic_visual_scoring_calculator#no1>div.dvsc_settings>table>tbody>tr>td>.dvsc_tab>p {
+    tab-size: 4;
+    margin: 0;
+    text-align: end;
+    white-space: nowrap;
+    word-break: keep-all;
+}
+
+
+div.dynamic_visual_scoring_calculator#no1>div.dvsc_settings>table>tbody>tr>td>.dvsc_tab>output.dvsc_category_score_display{
+    font-weight: bold;
+}
+
+div.dynamic_visual_scoring_calculator#no1>div.dvsc_output>.dvsc_output_svg {
+    height: 200px;
+    width: 200px;
+    position: relative;
+    overflow: visible;
+}
+
+div.dynamic_visual_scoring_calculator#no1>div.dvsc_output>svg.dvsc_output_svg * {
+    transition: all 0.4s ease-out;
+} div.dynamic_visual_scoring_calculator#no1>div.dvsc_output>svg.dvsc_output_svg>g#dvsc_total_score>circle#dvsc_score_indicator {
+    cx: 50;
+    cy: 50;
+    r: 40;
+    stroke-width: 4px;
+    stroke-linecap: round;
+    fill: none;
+}
+
+ div.dynamic_visual_scoring_calculator#no1>div.dvsc_output>svg.dvsc_output_svg>g#dvsc_total_score>text#dvsc_score_text {
+    text-anchor: middle;
+    font-family: 'Courier New', Courier, monospace;
+    font-size: 32px;
+    font-weight: bold;
+    dominant-baseline: middle;
+}
+
+div.dynamic_visual_scoring_calculator#no1>div.dvsc_output>svg.dvsc_output_svg>g#dvsc_fractions>circle.dvsc_fraction {
+    cx: 50;
+    cy: 50;
+    r: 46;
+    stroke-width: 3px;
+    stroke-linecap: round;
+    fill: none;
+    transform-origin: center;
+}
+
+div.dynamic_visual_scoring_calculator#no1>div.dvsc_output>svg.dvsc_output_svg>g#dvsc_fractions>circle.dvsc_fraction_bg {
+    opacity: 0.5;
+}
+
+div.dynamic_visual_scoring_calculator#no1>div.dvsc_output>svg.dvsc_output_svg>g#dvsc_fractions>circle.dvsc_fraction_bar {
+    opacity: 1;
+}
+
+
+div.dynamic_visual_scoring_calculator#no1>div.dvsc_output>svg.dvsc_output_svg>g#dvsc_fractions>text.category_text{
+    font-size: 8px;
+}
