+++ src/vanilla/index.html
@@ -1,6 +1,6 @@
 <!DOCTYPE html >
 <html lang="en">
-<body>   
+<body> 
         <div class="dynamic_visual_scoring_calculator">
             <div class="dvsc_settings">
                 <table>
